@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.9.0
+  sdk: ^3.10.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -58,6 +58,7 @@ dependencies:
   cached_network_image: ^3.4.1
   equatable: ^2.0.7
   restart_app: ^1.3.2
+  logger: ^2.6.2
 
 dev_dependencies:
   flutter_test:
