@@ -1,7 +1,6 @@
-import '../../../features/cart/domain/entities/cart_item_entity.dart';
 
 abstract class CartService {
-  Future<void> addItemToCart(CartItemEntity item);
+  Future<void> addItemToCart(String productId);
 
   Future<void> removeItemFromCart(String productId);
 
