@@ -7,6 +7,7 @@ import 'package:fruit_hub/core/services/database/cart_service.dart';
 import 'package:fruit_hub/features/cart/domain/entities/cart_item_entity.dart';
 import 'package:fruit_hub/features/cart/domain/use_cases/add_item_to_cart_use_case.dart';
 import 'package:fruit_hub/features/cart/domain/use_cases/get_cart_items_use_case.dart';
+import 'package:fruit_hub/features/cart/domain/use_cases/get_products_in_cart_use_case.dart';
 import 'package:fruit_hub/features/cart/domain/use_cases/remove_item_from_cart_use_case.dart';
 import 'package:fruit_hub/features/cart/domain/use_cases/update_item_quantity_use_case.dart';
 
@@ -16,21 +17,31 @@ class CartCubit extends Cubit<CartState> implements CartService {
   CartCubit(
     this._addItemToCartUseCase,
     this._removeItemFromCartUseCase,
-    this._getCartItemsUseCase,
+    this._getProductsInCart,
     this._updateItemQuantityUseCase,
+    this._getCartItemsUseCase,
   ) : super(CartInitial());
 
   final AddItemToCartUseCase _addItemToCartUseCase;
   final RemoveItemFromCartUseCase _removeItemFromCartUseCase;
-  final GetCartItemsUseCase _getCartItemsUseCase;
+  final GetProductsInCartUseCase _getProductsInCart;
   final UpdateItemQuantityUseCase _updateItemQuantityUseCase;
+  final GetCartItemsUseCase _getCartItemsUseCase;
+
+  List<Map<String, dynamic>> _cartItems = [];
+  List<CartItemEntity> _productsInCart = [];
+
+  bool isInCart(String productId) =>
+      _cartItems.any((item) => item['productId'] == productId);
 
   @override
-  Future<void> addItemToCart(CartItemEntity item) async {
-    final result = await _addItemToCartUseCase.call(item);
+  Future<void> addItemToCart(String productId) async {
+    emit(CartLoading(newItemAdded: true));
+    final result = await _addItemToCartUseCase.call(productId);
     switch (result) {
       case NetworkSuccess<void>():
-        await getCartItems(false);
+        _addToCartItemsLocal(productId);
+        await getProductsInCart(needLoading: false, newItemAdded: true);
       case NetworkFailure<void>():
         emit(CartFailure(getErrorMessage(result).tr()));
     }
@@ -38,93 +49,148 @@ class CartCubit extends Cubit<CartState> implements CartService {
 
   @override
   Future<void> removeItemFromCart(String productId) async {
+    emit(CartLoading(itemRemoved: true));
     final result = await _removeItemFromCartUseCase.call(productId);
     switch (result) {
       case NetworkSuccess<void>():
-        await getCartItems(false);
+        _removeFromCartItemsLocal(productId);
+        await getProductsInCart(needLoading: false, itemRemoved: true);
       case NetworkFailure<void>():
         emit(CartFailure(getErrorMessage(result).tr()));
     }
   }
 
-  Future<void> getCartItems([bool needLoading = true]) async {
+  Future<void> getCartItems() async {
+    emit(GetCartItemsLoading());
+    final result = await _getCartItemsUseCase.call();
+    switch (result) {
+      case NetworkSuccess<List<Map<String, dynamic>>>():
+        _cartItems = result.data!;
+        emit(GetCartItemsSuccess());
+      case NetworkFailure<List<Map<String, dynamic>>>():
+        emit(GetCartItemsFailure(getErrorMessage(result).tr()));
+    }
+  }
+
+  Future<void> getProductsInCart({
+    bool needLoading = true,
+    bool newItemAdded = false,
+    bool itemRemoved = false,
+  }) async {
     if (needLoading) {
       emit(CartLoading());
     }
-    final result = await _getCartItemsUseCase.call();
+    final result = await _getProductsInCart.call(_cartItems);
     switch (result) {
       case NetworkSuccess<List<CartItemEntity>>():
         final items = result.data!;
-        emit(
-          CartSuccess(
-            items: items,
-            totalItemCount: items.length,
-            totalPrice: _calculateTotalPrice(items),
-          ),
-        );
+        _productsInCart = items;
+        _emitCartSuccess(newItemAdded: newItemAdded, itemRemoved: itemRemoved);
       case NetworkFailure<List<CartItemEntity>>():
         emit(CartFailure(getErrorMessage(result).tr()));
     }
   }
 
   @override
-  Future<void> decrementItemQuantity(String productId) async {
-    final int? newQuantity = _getNewQuantity(
+  Future<void> incrementItemQuantity(String productId) async {
+    await _updateQuantityOptimistically(
       productId: productId,
-      increment: false,
+      isIncrement: true,
     );
-    if (newQuantity == null) return;
-    if (newQuantity == 0) {
-      await removeItemFromCart(productId);
-      return;
-    } else {
-      final result = await _updateItemQuantityUseCase.call(
-        productId: productId,
-        newQuantity: newQuantity,
-      );
-      switch (result) {
-        case NetworkSuccess<void>():
-          await getCartItems(false);
-        case NetworkFailure<void>():
-          emit(CartFailure(getErrorMessage(result).tr()));
-      }
-    }
   }
 
   @override
-  Future<void> incrementItemQuantity(String productId) async {
-    final int? newQuantity = _getNewQuantity(productId: productId);
-    if (newQuantity == null) return;
+  Future<void> decrementItemQuantity(String productId) async {
+    await _updateQuantityOptimistically(
+      productId: productId,
+      isIncrement: false,
+    );
+  }
+
+  Future<void> _updateQuantityOptimistically({
+    required String productId,
+    required bool isIncrement,
+  }) async {
+    final index = _productsInCart.indexWhere(
+      (e) => e.fruitEntity.code == productId,
+    );
+    if (index == -1) return;
+
+    final currentItem = _productsInCart[index];
+    final int oldQuantity = currentItem.quantity;
+    final int newQuantity = isIncrement ? oldQuantity + 1 : oldQuantity - 1;
+
+    if (newQuantity == 0) {
+      await removeItemFromCart(productId);
+      return;
+    }
+
+    _updateLocalListQuantity(productId, newQuantity);
+    _emitCartSuccess();
+
     final result = await _updateItemQuantityUseCase.call(
       productId: productId,
       newQuantity: newQuantity,
     );
+
     switch (result) {
       case NetworkSuccess<void>():
-        await getCartItems(false);
+        break;
+
       case NetworkFailure<void>():
-        emit(CartFailure(getErrorMessage(result).tr()));
+        _updateLocalListQuantity(productId, oldQuantity);
+        _emitCartSuccess();
+      // emit(CartFailure(getErrorMessage(result).tr()));
     }
   }
 
   // -----------------------------------------------
 
+  void _removeFromCartItemsLocal(String productId) =>
+      _cartItems.removeWhere((item) => item['fruitCode'] == productId);
+
   double _calculateTotalPrice(List<CartItemEntity> items) => items
       .map((e) => e.totalPrice)
       .fold(0, (value, element) => value + element);
 
-  int? _getNewQuantity({required String productId, bool increment = true}) {
-    final currentState = state;
-    if (currentState is! CartSuccess) {
-      return null;
-    }
-    try {
-      final item = currentState.items.firstWhere(
-        (i) => i.fruitEntity.code == productId,
+  void _emitCartSuccess({bool newItemAdded = false, bool itemRemoved = false}) {
+    emit(
+      CartSuccess(
+        items: List.from(_productsInCart),
+        totalItemCount: _productsInCart.length,
+        totalPrice: _calculateTotalPrice(_productsInCart),
+        newItemAdded: newItemAdded,
+        itemRemoved: itemRemoved,
+      ),
+    );
+  }
+
+  void _updateLocalListQuantity(String productId, int quantity) {
+    final productIndex = _productsInCart.indexWhere(
+      (e) => e.fruitEntity.code == productId,
+    );
+    if (productIndex != -1) {
+      final oldItem = _productsInCart[productIndex];
+      _productsInCart[productIndex] = CartItemEntity(
+        fruitEntity: oldItem.fruitEntity,
+        quantity: quantity,
       );
-      return increment ? item.quantity + 1 : item.quantity - 1;
-    } catch (e) {
-      return null;
+      _productsInCart[productIndex] = oldItem.copyWith(quantity: quantity);
+    }
+    final cartIndex = _cartItems.indexWhere((e) => e['fruitCode'] == productId);
+    if (cartIndex != -1) {
+      _cartItems[cartIndex]['quantity'] = quantity;
+    }
+  }
+
+  void _addToCartItemsLocal(String productId) {
+    int index = _cartItems.indexWhere(
+      (element) => element['fruitCode'] == productId,
+    );
+    if (index != -1) {
+      _cartItems[index]['quantity']++;
+    } else {
+      _cartItems.add({'fruitCode': productId, 'quantity': 1});
     }
   }
 }
